# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BYUconnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export interface Building {
  id: number;
  name: string;
  abbreviation: string;
  latitude: number;
  longitude: number;
  address: string;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  icon: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  buildingId: number;
  buildingName: string;
  roomNumber: string;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  clubId: number;
  clubName: string;
  capacity: number;
  reservedCount: number;
  hasFood: boolean;
  coverImageUrl?: string | null;
  tags: string[];
  isSaved: boolean;
  isReserved: boolean;
  latitude: number;
  longitude: number;
}

export type EventDetail = Event & {
  clubAvatar?: string;
  clubDescription?: string;
};

export interface CreateEventRequest {
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  buildingId: number;
  roomNumber: string;
  categoryId: number;
  clubId: number;
  capacity: number;
  hasFood: boolean;
  tags: string[];
}

export interface SaveToggleResponse {
  saved: boolean;
}

export interface ReservationToggleResponse {
  reserved: boolean;
  reservedCount: number;
}

export interface Club {
  id: number;
  name: string;
  description: string;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  memberCount: number;
  avatarInitials: string;
  avatarColor: string;
  coverImageUrl?: string | null;
  contactEmail: string;
  isMember: boolean;
}

export interface Announcement {
  id: number;
  clubId: number;
  title: string;
  body: string;
  createdAt: string;
}

export type ClubDetail = Club & {
  upcomingEvents?: Event[];
  announcements?: Announcement[];
};

export interface JoinToggleResponse {
  joined: boolean;
  memberCount: number;
}

export interface UserProfile {
  user: User;
  savedEvents: Event[];
  reservations: Event[];
  savedCount: number;
  reservationsCount: number;
}

export interface SearchResults {
  events: Event[];
  clubs: Club[];
}

export type GetEventsParams = {
  categoryId?: number;
  buildingId?: number;
  clubId?: number;
  search?: string;
};

export type GetClubsParams = {
  categoryId?: number;
  search?: string;
  myClubs?: boolean;
};

export type SearchParams = {
  q: string;
};