# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BYUconnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clubs:
    get:
      operationId: getClubs
      tags: [clubs]
      summary: Get clubs
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: myClubs
          schema:
            type: boolean
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
  /clubs/{id}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get club by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{id}/join:
    post:
      operationId: joinClub
      tags: [clubs]
      summary: Join/leave club
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Membership toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinToggleResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required: [email, password, firstName, lastName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
      required: [id, email, firstName, lastName, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        buildingId:
          type: integer
        buildingName:
          type: string
        roomNumber:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        clubId:
          type: integer
        clubName:
          type: string
        capacity:
          type: integer
        reservedCount:
          type: integer
        hasFood:
          type: boolean
        coverImageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isSaved:
          type: boolean
        isReserved:
          type: boolean
        latitude:
          type: number
        longitude:
          type: number
      required: [id, title, description, startTime, endTime, buildingId, buildingName, roomNumber, categoryId, categoryName, categoryColor, clubId, clubName, capacity, reservedCount, hasFood, tags, isSaved, isReserved, latitude, longitude]
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        memberCount:
          type: integer
        avatarInitials:
          type: string
        avatarColor:
          type: string
        coverImageUrl:
          type: string
          nullable: true
        contactEmail:
          type: string
        isMember:
          type: boolean
      required: [id, name, description, categoryId, categoryName, categoryColor, memberCount, avatarInitials, avatarColor, contactEmail, isMember]
    Announcement:
      type: object
      properties:
        id:
          type: integer
        clubId:
          type: integer
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
      required: [id, clubId, title, body, createdAt]
    ClubDetail:
      allOf:
        - $ref: "#/components/schemas/Club"
        - type: object
          properties:
            upcomingEvents:
              type: array
              items:
                $ref: "#/components/schemas/Event"
            announcements:
              type: array
              items:
                $ref: "#/components/schemas/Announcement"
    JoinToggleResponse:
      type: object
      properties:
        joined:
          type: boolean
        memberCount:
          type: integer
      required: [joined, memberCount]
...