# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BYUconnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  clubId: number;
  title: string;
  body: string;
  createdAt: string;
}