# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BYUconnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: number;
  name: string;
  description: string;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  memberCount: number;
  avatarInitials: string;
  avatarColor: string;
  coverImageUrl?: string | null;
  contactEmail: string;
  isMember: boolean;
}