# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BYUconnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}